// types/couponDataType.ts
export interface CouponDataType {
    "id":number
    "coupon_id": number
    "coupon_logo_img":string
    "coupon_img":string
    "coupon_value_img":string
    "coupon_name": string
    "coupon_desc": string
    "reg_date": string
    "coupon_reg_date":string
    "coupon_ex_date": string
    "coupon_type": string
    "available_place": string
    "coupon_value": number
}